backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
dj-database-url==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
drf-spectacular==0.26.5
Pillow==10.1.0
requests==2.31.0
phonenumbers==8.13.26
qrcode[pil]==7.4.2